style: polish auth screens
