+++ src/app/(with-bg)/login/page.tsx
@@ -1,0 +1,58 @@
+'use client';
+
+import { useState, FormEvent } from 'react';
+import { useRouter } from 'next/navigation';
+import { supabase } from '@/lib/supabaseClient';
+import { cn, glassClasses } from '@/lib/utils';
+
+export default function LoginPage() {
+  const router = useRouter();
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [error, setError] = useState<string | null>(null);
+
+  async function handleSubmit(e: FormEvent) {
+    e.preventDefault();
+    setError(null);
+    const { error } = await supabase.auth.signInWithPassword({ email, password });
+    if (!error) {
+      router.push('/');
+    } else {
+      setError(error.message);
+    }
+  }
+
+  return (
+    <div className="min-h-screen flex items-center justify-center">
+      <form
+        onSubmit={handleSubmit}
+        className={cn(glassClasses, 'p-4 max-w-md w-full flex flex-col gap-4')}
+      >
+        <h1 className="text-xl font-bold text-center">Login</h1>
+        <input
+          type="email"
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+          placeholder="Email"
+          className="border px-3 py-2 rounded"
+          required
+        />
+        <input
+          type="password"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+          placeholder="Password"
+          className="border px-3 py-2 rounded"
+          required
+        />
+        {error && <p className="text-red-600 text-sm">{error}</p>}
+        <button
+          type="submit"
+          className="bg-blue-600 text-white rounded px-4 py-2"
+        >
+          Login
+        </button>
+      </form>
+    </div>
+  );
+}
